backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
groq
requests
flask
flask-cors
python-multipart
httpx